added example module
