+++ src/main/java/io/apollo/utils/GLRenderer.java
@@ -19,6 +19,7 @@
 
 package io.apollo.utils;
 
+import io.apollo.Apollo;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.Gui;
 import net.minecraft.client.gui.ScaledResolution;
